renamed getKeywordCount to keywordCount
